Avoid invalid uv_write after resume
After resuming an IO actor, specifically a TCP actor in this case, the
fd we had is invalid. We recently changed so that the __resume__ method
would set fd = -1 but apparently that's not enough to avoid libuv from
segfaulting. I figured it would just not run uv_write, perhaps returning
with an error, when fed an invalid fd but nope, it segfaults. So now we
return early from these functions if we got an invalid fd.