Add aid() for getting actor id
str(), repr() or similar on actors isn't working. In order to get some
unique identifier on an actor for the purpose of putting in a dict or
similar I'm adding this aid() method which returns the actor id as an
integer.

Unfortunately it has to be made per actor type since we don't have any
form of inheritance here... I think. So adding it where it is needed,
this is really just a band-aid in the short term (famous last words,
heh).