Do garbage collection using libgc
This adds garbage collection to the Acton run time system through the
use of libgc. We use link time redirection of malloc and friends in
order to catch all occurrences. Thread creation is hooked by including
gc.h in RTS and the DB client (which are the two places where we create
threads). free() is turned into a nop.

An earlier attempt was made to change our source code from using malloc
& friend to GC_MALLOC & friends but as it is exceedingly difficult to
find all mallocs, it regularly produced programs that segfaulted. libuv
and yyjson allows using a custom malloc but other libraries, like
utf8proc, doesn't appear to have functionality for specifying which
malloc to use and so it becomes rather difficult. Thus, this approach
was abandoned in preference of link time redirection of `malloc`.

Thread creation is hooked in source code as this is the only way that
works on MacOS. GC_USE_DLOPEN_WRAP allows wrapping dlopen so that on
Linux, hooks are installed at run time when the application is
dynamically linked with libpthreads. This facility does not exist on
MacOS which is why we are instead hooking it via source (by setting
GC_THREADS 1 and including gc.h). It ought to be more stable setting
this in source too. This is however untested, libgc even has an error in
gcconfig.h defined for this, so there's some uncertainty if there might
be bugs. It seems to work though.

Unlike other external libraries we depend on, we do not put libgc into
libActonDeps as we use libActonDeps both for Acton applications as well
as for actondb. We do not want actondb to use libgc and thus libgc must
be kept separate.

libgc allows control via environment variables, most notably export
GC_DONT_GC=1 will disable the GC. This can be useful in case we suspect
the GC causes some problems. Once we've gained confidence with the GC,
we should disable this at compile time by setting -DNO_GETENV.

I've tested this primarily with the perf/ring program, which is a form
of benchmarking program that sends messages between actors as fast as
possible. Before this patch, on my laptop, ring consumes around 12GB of
RAM after running for about 1 minute sustaining around 800k messages per
second. With libgc, memory consumption stays constant at 4.5MB while
maintaining roughly the same message rate. I haven't collected a lot of
statistics and the variation between test rounds is probably around 10%.
I cannot see a clear pattern between using GC and not on this
performance.

As far as I understand, libgc does pause all threads every now and then
in order to collect a correct view of registers and similar. This means
that concurrency likely will be hit harder in Acton than the kind of
single threaded / low concurrency applications used in the above
mentioned benchmarking. ring is actually quite "single threaded" in that
there is only one outstanding message in total being passed between the
actors so we might not see those effects clearly.

The longer term plan is obviously to implement our own GC that does
things the Acton way and will be way more efficient in various but until
then... this will do, I think :)