Improve test_tcp_server test
I've added some more debug printouts, in particular we now display the
error from the TCPIPListener, which should normally be "resume" when the
system resumes but could perhaps be something else.

Timeouts have been increased, although I'm not sure if this has a real
effect or not.

In my testing, these changes to the test scripts drops the error rate
for this test on Ubuntu 20.04 from 13% to ~1.5% which is in line with
other platforms.