Combine all external libs in one .a archive
We depend on multiple external libraries and in order to produce a known
and consistent result, we include those libraries in the Acton
distribution.

When actonc compiles, since we produce C code, we need to link that with
the libraries of our external dependencies which means the acton
compiler needs to have an exhaustive list of all libraries. This
includes libraries required by builtins and by stdlib. It is not
desirable to have to maintain such a list in multiple places (both
Makefile where we copy files and in compiler for correct linking).

We now combine all .a libraries into one big .a archive which we can
then link with and so it's up to cc/ld to figure things out. Thus we've
remove the definition of all these libraries from actonc!