Build all library deps from source
Our strategy thus far has been to rely on the package manager of system
on which we are being built. On Linux (Debian) we've used apt to install
packages and then copy the .a archives from those packages into the
Acton distribution, which means we get a library built with some set of
options which we are unable to influence. On Mac we do the similar thing
but using Homebrew.

Not all libraries are available in apt or brew, so there is a point
where we have to start compile things from source. We now do and beyond
being able to bring in libraries that are not available in apt/brew, we
also gain better control over what compilation flags are used, so for
example, we include libxml2 but without z, lzma or python support,
making it a much smaller library with no other dependencies.

Our build time naturally increases due to this but it's not too bad,
perhaps a minute or two on a fast machine.

I have opted to use neither git submodules nor git subtree for the
dependencies simply because it is not needed. subtree is a no go for
things like libuuid, which is part of util-linux, which weighs in at
250MB!!

Each dependency is pulled in via git over https and we check out a
particular tag or commit for each, so they are effectively pinned. I
have not spent much time considering the tag / commit to use, rather
just picked like the latest and it seems to work.

Make target prerequisites are not correctly setup, so we will not
correctly invalidate these dependencies nor make sure they are up to
date. Pretty much the only check is that lib/libActonDeps.a exists. If
it does not, we will try to build all the libraries, whereas if it does
exist, we will do nothing. Thus, caching works but cache invalidation
does not. It's a manual process for now. make clean will clean out the
libs but leave the repositories, making it possible to recompile with
other flags without having to redownload (good for airplane mode!).

Also added libxml2! Unlike the version in debian, we compile ours
without lzma, z or any other dep making it smaller and faster to
compile. It has no external dependencies!