Bundle and use zig
This bundles zig into the Acton distribution and uses it to build the
Acton distribution itself and later actonc uses zig for compiling the C
version of an Acton program. zig zig zig!

We use zig supercool cross-compilation to target glibc 2.28 which should
allow us to run on much older distributions, very cool!