Add & use actonc --always-build
--always-build skips the up-to-date check in actonc and always builds,
which is necessary for correct tests since we otherwise risk relying on
outdated results. The up-to-date check only inspects source and output
files on disk. For example, if a .ty file is newer than its source .act
file. If actonc itself has changed and updated the format of the .ty
file, we won't catch this. --always-build is thus the only safe option
for testing.