Move libxml2 headers in place
libxml2 seems to insist on installing in include/libxml2 but then must
be accessible as libxml/foo, i.e. we must place include/libxml2 in our
include path with -I... This includes putting it in actonc and that's
just rather ugly when all we need it for is our stdlib compilation.
Maybe we could add in some way to add compile time flags for an acton
project but until we have that I prefer not to pollute the compiler and
thus I've simply moved the libxml dir over so it's directly accessible
in our include path.