Skip dbgsym debs
Seems they are no longer generated when we compiled with zig!? Not even
sure what were in them given that we're a haskell compiler and some
libraries which are later linked into the acton compiled programs...
does dbgsym separate out debug symbols even in libActonDeps and similar?
Seems unlikely but if it is then it's not something we want either way.