Print backtrace on SIGILL & SIGSEGV
This is quite a magical hack. We capture the SIGILL & SIGSEGV signals
and run the debugger on ourselves to print a backtrace in case we hit
SIGILL or SIGSEGV. gdb is the best pretty printer there is of the stack
so while it's a little dirty not to use a library, there's just no way
we could achieve like it any other way. While it is quite trivial to
produce a stack trace, it will only include function names. gdb output
also includes source code files and line numbers plus local variables -
it's quite some magic.

Once we're done printing the stack trace, we kill ourselves with the
original signal (after removing our special signal handlers), meaning we
can still dump core if the user has configured his system accordingly.

There are currently some problems using lldb on macos, while it works
fine locally, the tests hang in CI so this is left commented out for
now. Thus we only support using gdb, which is predominantly used on
Linux distributions. In case gdb is not available, we simply don't
print the backtrace, just a brief error.