Flip bigint example to int test case
Changed around the example of bigint to be a test case for
our (big)ints! Choose not to call it bigint since our normal integers
are bigints but we don't be calling them that. This can certainly be
expanded upon but it's a start.

Also placed in tests/builtin_auto/int.act as it should be!