Add --auto-stub to actonc
This enables the automatic detection of stub mode compilation.
Previously it was the default to always do stub mode detection, but this
is problematic if there happens to be .c files with the same name as our
.act file in the same directory but it's entirely unrelated, i.e. not an
Acton C module implementation.

This is the case with the Programming Languages Benchmark where all test
files are in the same directory and named by a sequence number, like
1.act is the first Acton solution to a particular problem and 1.c is the
C solution to the same problem.

It is saner and safer to not do automatic stub compilation per default.
For some places, like our stdlib, we now have to provide --auto-stub,
which feels just fine!