actondb gossip listen on any adress & avoid UB
The reason I'm doing this here and now is because the bcopy code
triggers undefined behavior on macos.

With that said, I think this change is conceptually correct, regardless
of UB. h_addr_list[0] seems to be 127.0.0.1 on my machine and I assume
that it always is, thus the intention is to listen on 127.0.0.1 - but
that's not what we actually want for actondb. The gossip protocol needs
to run between actondb instances on different machines so it must listen
on any (0.0.0.0).