Lower test timeouts
We should always reach a no quorum error in 10 seconds. We set the test
timeout to 2 times that plus a second of margin so we really should be
able to see 2 such messages. It's just unnecessary waiting any longer
and we could probably consider lowering it even further (maybe changing
no quorum error timeout too?). This should speed up failure cases in
tests a bit (if we run a lot of tests and expect some level of failures).