Add actonc --cc for what CC to use & add --verbose
This makes it possible to pick what C compiler is used when compiling an
Acton program. While the default is still zig cc I have noticed that we
get poor performance in certain situations and it makes things a lot
easier to debug if we can easily switch CC at run time rather than
having to recompile actonc.

In the few cases where I have tested this, it has been enough to only
change CC during the final acton program compilation, i.e. the acton
system, including RTS, builtins and stdlib is still built with zig cc.
There might be other cases where we would need to recompile that as
well, which is then an entirely different matter.

Also, some of the options we pass to zig are known to be zig specific,
so those are now only thrown in for the default of zig cc.

When we compile using zig cc, we now run with --verbose which might help
debug some scenarios.