Add install instructions & remove GC comment
The note on lack of GC is outdated since we now have a GC!

While the web page install instructions are nice, I have regretted
removing the install instructions from the readme since I just think
it's nice that the instructions, in their very briefest form, are
available right in the repo. Makes it easier to grasp how simple the
install process is.