Add Containerfile & build images in CI
Add CI job to build container images and publish them on GHCR. We build
a multi-arch image for arm64 and amd64 (x86_64) tagged with the normal
tip version, though we've removed the seconds from that version
timestamp, for example `ghcr.io/actonlang/acton:0.26.0.20260109.12.43`.
There's a `tip` alias pointing to the latest one (or really the last CI
job run, so reruns of old jobs could slightly mess this up but it'll
mostly be OK). In order to align the tip version number for the multiple
jobs required to produce the multi-arch image, we have restructured CI
jobs so there is a new build-version job whose only job is to come up
with the minute-precision BUILD_TIME which is then propagated to other
jobs, keeping tip versions aligned across debs and images. :latest
points to the last numbered release.

There are actually two ways of building the container images. What we
use in CI is based on first producing .deb files and then installing
those, this is defined in `Containerfile.deb`. The second variant, which
is really main "main" version and normally encouraged at least for local
build and use of container images, is in `Containerfile` - it copies the
dist/ directory into the container and thus shortens the time to build
the container considerably (building debs is slower), thus allowing
faster iteration.