Handle signal interrupts in rts.sleep
Calling usleep isn't signal safe and since our GC is using signals
rather extensively to pause threads, we ended up with a fairly broken
rts.sleep. By switching to nanosleep we not only tolerate signals but
can also resume sleeping to the correct amount as nanosleep will return
the remaining time that it didn't sleep.

Also added in a acton.rts.gc() function that will explicitly invoke the
GC since this was fairly useful to trigger bugs during testing. There's
an "Interruptor" actor in the test of acton.rts.sleep which was helpful
during development but as it seems we have some unknown deadlock!? I am
not leaving it enabled, though just commented out so we can investigate
further.