Allow overriding number of worker threads
We've allowed setting the number of worker threads for a long time but
we have effectively not honored numbers < 4. This wasn't really
deliberate, on machines with few cores we run a minimum of 4 worker
threads regardless because we think that's better than say 1 or 2
threads. With 4 threads, the OS will task switch between them and allow
progress on multiple actors concurrently, which can be better than being
blocked on a single actor. But it also meant overriding even when the
user had explicitly configured the number of worker threads.

This is now fixed so we always honor the wish of the user.