Bundle timex.h and copy in on build
Rather than relying on timex.h existing at some particular path, we now
bundle timex.h ourselves and copy it in at build time. It seems that on
Macos 13, or perhaps due to some other circumstances, timex.h might be
in a different location. It's just easier to bundle it ourselves than
try to find it across a number of different paths.

The final solution would be to have it included in upstream zig instead.
We have initiated this work.