Filter binTasks
When the --root argument is not provided binTasks contains all modules
we know about, like all project source files, which means we invoke the
buildExecutable target for all files and if they don't contain a 'main'
actor, the file is skipped.

We now filter the binTasks, or rather preBinTasks to binTasks a little
earlier. This will simplify other changes to our build system later on
when we now have a cleaner definition of binTasks.