Add http module
Adds a rather basic http & https module. There are many things to
improve here, but it basically works. We can adjust and fix over time.
Having a http module in stdlib is a big and important step!

There are some tests but nothing is run automatically, primarily because
I do not want to introduce a dependency on having Internet connectivity.
We should test against ourself though! And we should run unit tests..