Run CC process assuming utf8 on stdout/err
With readCreateProcessWithExitCode() as we previously used it, we got an
error while compiling test/builtin_auto/builtin_functions_test.act:

  ...
  actonc: fd:7: hGetContents: invalid argument (invalid byte sequence)

The problem was noticed in CI for a new branch, which does not modify
these parts of the compiler, yet fails on an existing test. It is
reproducible on a Linux machine when LANG is not set (CI does not set
LANG), so the system defaults to LANG=C (I think). The hGetContents is
not called directly from our code (we enforce encoding where we do), but
it must be called from within readCreateProcessWithExitCode() from the
standard System.Process library. We have no way of modifying it. It's
puzzling because the output of the commands run do not appear to contain
anything but US ASCII, so it's unclear how we can get an invalid byte
sequence. And why is this only happening in a branch with seemingly
unrelated changes?

I refactored the code to use createProcess and async reading of stdout
and stderr (to avoid blocking), which allows more direct control over
the stdout & stderr pipes.

Running CC is also moved to a separate function that can be called from
the two places where we run CC. Less code!

Also, while we now enforce encoding, this is just enforcing our assumed
encoding when reading. The processes we run could be outputting
something else. I presume they honor locale, so it might in fact not be
UTF-8 now. But can we influence that in a portable way? I don't see how,
setting LANG=en_US.UTF-8 won't work on a machine with sv_SE.UTF-8 and it
won't work on another OS?