Include platform in deps archive name
libActonDeps.a is the archive we build containing all our external
library dependencies, like libuv, libargp etc. In order to support
multiple platform (for cross-compilation) we now start including the
platform name in the file name.

Many other steps are necessary to reach cross-compilation, but this is
one step closer.