Add deps sum
For each dependency library we have, we also have a git ref specified
which is what we check out. It's a form of rather explicit version pinning.

By taking all those refs and checksumming them, we get a combined
checksum for all our dependencies. This in turn can be used to determine
if two libActonDeps have the same content or not, which is what we need
in order to cache it.

- show-deps-sum is a new target that simply prints this checksum.
- builds-deps is a target to build libActonDeps for the current platform