Revamp deps build & use deps dl as cache
This revamps building deps so we now produce a single tar.bz2 ball
containing relevant libraries and headers. Makes it easier to handle
rather than multiple independent files.

There is now some support, though relatively untested, for also
downloading an existing tar ball, extracting it and using those files.
It is possible to override by setting ALWAYS_BUILD=true, which means we
will not download and force local build. This should also be idempotent,
so we do not attempt to download the deps archive multiple times, thus
even without internet connectivity, we won't attempt to run curl (which
would fail).

The dist/include directory has been moved to dist/inc, this is because
Homebrew automatically recognizes this as a standard package directory,
it thinks we are a library and want to expose our header files, which is
not what we want. To avoid the automagic smart detection, we've simply
moved it. All of this will be revamped in the future anyway. It hasn't
caused a problem previously because we were very selective about what
headers to place there. Now, for simplicity, we include all header files
from our deps, which triggers some form of error in CI for the brew test
build.