Build a portable builder!
One pin down the rabbit hole! Switched from using the `zig build`
frontend, which under the hood just builds the builder program based on
our build.zig. It's really build_runner.zig, which in turn imports
build.zig. We now do that step manually and are then able to inject our
own compilation arguments, like which arch to compile for and should
thus get a portable executable that is built for the base CPU feature
set without all the bells and whistles of the native CPU of the build
machine!

Default to use the zig build!