Move builtin,rts,stdlib -> base Acton project
Join builtin, rts & stdlib into "base", which is structured like an
Acton project and can thus be compiled with `actonc build`.

The entry point for builtin is moved from builtin.c to __builtin__.c,
which is naturally rendered by actonc and then compiled.

ActonCompiler compileTasks is modified to check if __builtin__ is among
the current list of CompileTasks and runs that first, which makes it
possible to compile a project that contains __builtin__ and other
modules. Also reinit env and honour the path, either a __builtin__ in
syspath or in projpath (if it was part of compileTasks). Although not
tested, this should also make it possible in the future to use a
customized base package.

We now make includes relative to the project root, for example
types/foo.h -> out/types/foo.h, which clears up a few things. The base
package in the system distribution is now implicitly included and is
included in the way we imagine other package dependencies will be
included, so the relative paths within a project are always the same for
the local package, dependency packages and base (which is an implicit
dependency).

The entire base project is also included in the acton distribution, i.e.
we ship the C source code, making it possible in the future, to
recompile base, for example with different compile flags or targeting a
different architecture.

checkUpToDate has been fixed, now comparing against the generated C file
instead of the o, since it might not exist and we leave it to Zig to
compare C to O.

The support for custom Makefiles has been ditched. It was used to
compile math and numpy in stdlib but due to the restructuring, it is no
longer necessary. These modules are stub compiled, which we detect by
the presence of a .c file, like src/math.c next to src/math.act. We copy
the .c and .h files to out/types/, so the output from actonc looks the
same as if our CodeGen had written the C file. To our C/Zig builder, the
input C source files look the same. Given the changed include
paths (relative to project root), include statements have been updated.