Skip dh_strip
Debian runs strip per default on stuff when building a debian package.
Strip doesn't like the absolute paths we have in our archives. Since we
are building with zig, we might as well ask zig to do the stripping for
us and so we can skip the dh_strip stage altogether.

I think that dh_strip is responsible for extracting debug symbols, so
that might be something we loose right now, but I think that's ok. It
was always a little weird doing separation of debug symbols for the
acton package anyway, it's a development environment, so it might as
well come with debug symbols. It's more interesting what to do with
debug symbols for the output of actonc, not the acton debian package!