Directly include arc4random
arc4random comes standard on BSD systems. For Linux we require libbsd,
but since we only really want arc4random out of everything that libbsd
offers, and libbsd is quite iffy to compile, I'm opting here to directly
include a few of the arc4random related files instead of everything.
Let's see if we can get this to work reliably?