Use zigs ar to avoid platform inconsistencies
To build libActonDeps, which consists of multiple other arhives, we
first build those archives individually, then extract them to get the .o
files, the create libActonDeps from that. We use the `ar` tool for this
but it behaves differently on different platforms. That hasn't been an
issue in the past, but now that libyyjson is built using zig and not the
platform cc, it breaks down on MacOS. zig adds object files with
absolute paths into the archive and MacOS ar honors the absolute paths,
so when we extract libyyjson.a the result doesn't go in
lib_deps/libyyjson like we want, but is written back to like
$HOME/acton/deps/libyyjson/zig-cache/19f13e03...a9f/libyyjson.o so when
we build libActonDeps.a there libyyjson.o isn't included. GNU ar on
Linux doesn't let one extract files outside of the current directory, so
paths are rewritten to become relative and thus it worked.

Switching to zig ar, since it has the same behavior as GNU ar and will
thus work on MacOS. Overall it's a good idea to rely on the zig provided
variants since they are consistent across platforms.

Makefiles variables have been rewritten to make things easier.
$(DIST_ZIG) is what to depend on for install of zig itself, and we run
it from dist/zig. $(ZIG) points to the zig executable itself with an
absolute path so it works regardless if we enter into subdirectories or
not.