Downgrade arc4random -> rand() for portability
arc4random is in BSD but not available on Linux. We've been relying on
libbsd to get arc4random on Linux.

By using rand() instead, we don't need libbsd and can get rid of it and
in turn libmd which was a prerequisite of libbsd. These two libraries
have been rather tricky to compile, so removing them simplifies building
libActonDeps quite a lot!