Target glibc 2.27
Tests are failing on Ubuntu 18.04 where we are trying to use acton,
which is built on a recent Debian, to compile and run an Acton program.
It complains about not having GLIBC_2.28. Ubuntu 18.04 has glibc 2.27 so
it's not surprising that it fails and we shouldn't expect that to have worked.

I don't understand how this could ever have worked before when we
compiled libuv using the regular makefile but with zig cc targetting
x86_64-linux-gnu.2.28. Why didn't that break? Somehow switching to
build.zig breaks it?

Compiling using glibc 2.27 actually seems fine, so doing that now. We
used to have some form of dependency that did not work on anything under
2.28 but I don't remember what that was and it doesn't seem to be a
problem anymore.