Remove failing brew-macos job
The brew-macos job is failing more times than not. I don't have the time
or energy to fix it now, so option to remove as all failures are
becoming a nuisance.

I originally added the brew-macos job as we had failures once we pushed
a new release to the homebrew-acton repo. With rather few releases and
far between, the brew build problems only dawned on us once we made a
release. Having the Formula committed locally and testing it
continuously was an attempt in increasing the likelihood of finding
problems earlier so we could fix them. It did work well and did catch a
bunch of problems but something is different now, presumable in brew
itself, so that our local test no longer works. It was quite hacked up
and not a "supported mode of operation" in / by brew.

In the longer term I think we should add this job back, but I simply
don't have the energy to try and figure this stuff out now.

There is one big difference now compared to before when we had lots of
problems with brew builds (and which this test helped find), which is
related to dependencies. Before we had a number of external library
dependencies added as brew dependencies, so we let brew install static
libraries that we then copied into our distribution. Whenever we added
or removed a dependency, the brew Formula had to stay in sync (and we
often forget!). Now we do not have any externl dependencies like that,
instead we download and build external libs from source, so the risk to
run out of sync in this way is 0. There are some other build
dependencies and other parts of the Formula that might cause problems of
course, but the historically biggest contributor to problems is now, by
design, gone forever!

I left the Formula itself in this repo. No point in removing it...