Add --target arg to actonc
Allows one to override the target using a target triple, like
x86_64-linux-gnu.2.27 or x86_64-macos-none.

When we are targetting the default architecture, we can use the prebuilt
libraries we have instead of building from scratch. This practically
solves the problem of ephemeral build-caches for single .act file
compilation since the common case is to build for the default target. If
someone wants cross-compilation, I think it's acceptable that it takes
longer.