Clean up dist build
Now that we are using zig, new possibilities have opened up. We no
longer need to stage things as much, we can copy in sources directly to
dist/ and then let zig build it there using an out-of-tree build,
installing into dist/depsout

depsout is a bit of a lie since it's not just deps, but no matter?

Also fixes actual Zig caching for CI. Variable should be called
ZIG_LOCAL_CACHE_DIR and not ZIG_CACHE_DIR...