Ditch libActonDeps, minimize dist
In order to simplify compilation and decouple the compiler from the list
of C libraries, we bunched them all up in one large archive called
libActonDeps.a. Recently it was moved to include the platform name, like
libActonDeps-x86_64-linux.a.

With the move towards using zig build and how the entire build process,
both what's called from actonc as well as how we put together the Acton
distribution itself, the need for having libActonDeps is smaller. We
accept the list of libraries in multiple places. Since our dist building
step now naturally installs the .o files in dist/depsout, we make use of
them directly, so also building libActonDeps is just unnecessary. It has
now been removed.

The classic C compilation (not using zig) has also been fixed to use the
correct paths.

dist/builtin is gone since it is already in dist/base/builtin

dist/zig is now stripped of docs, libcxx and windows headers which
reduces it's size by over 100MB.