Add .close() for TCP client actor
This makes it possible to explicitly shut down a TCP client connection,
which should clean up everything in the libuv loop.

If actors are garbage collected, there is likely remaining socket cruft
left in the libuv loop which we do not currently have any means of
cleaning up. At least this gives us the ability to explicitly shut
things down cleanly from within the application.

The .close() method takes callback which is called when the connection
has been closed.