Add acton.rts.io_handles to get IO info
This returns the handles currently registered in the libuv I/O loop. It
is keyed on the handle (a pointer) and the information is a tuple
consisting of the handle type and the actor it is associated with, if
available. For the TCP actors, the actor is the relevant TCP actor.

Since we run a IO loop per worker thread, this is currently hard to use!
But it is useful with --rts-wthreads=1. We would need explicit actor
affinity to make it more useful.