Test reconnect & close
The TCP test now attempts a reconnect after establishing the initial
connection and sending a PING. Before exit, it verifies the expected
number of TCP handlers are registered in the I/O subsystem, to make sure
we've properly cleaned up after ourselves. This isn't entirely reliable
in CI right now because we run with multiple worker threads and we can
only retrieve the handlers for the current worker thread. This should be
fixed. For local testing, it can reliable check by running with
--rts-wthreads=1