Use relative path for __FILE__ / in logging
Our RTS log includes the path to the source file. Zig uses absolute
paths, so we have gotten a very long absolute path. By using
-ffile-prefix-map we can rewrite it to be a relative one.

We back out one directory from the buildroot, so that we not just see
the local path in a project but also which project directory. This isn't
a guarantee for global uniqueness, but it's probably good enough,
striking a balance between correctness and readability.