Improve support for alternative output from actonc
Using alternative output, like --sigs for printing signatures or --cgen,
we now suppress various log output by effectively assuming the same as
--quiet.

A bug meant that we would hit an error when running for a single .act
file and there is no actor called main. This is now fixed, so we still
perform the final compilation even when there is no main actor. The full
compilation can thus be tested.

If alternative output is enabled, like --cgen or similar, we exit early
and avoid the final compilation.