Rename Auth to Caps
In order to better reflect the mental model and intuition we want to
instill in users of Acton, we are renaming the Auth objects to Cap. With
a reference to a capability, we can use that capability. This aligns
with the existing security model for the internal actor domain where an
actor must hold a reference to another actor in order to call its
method. We want to extend that mental model for all external things in
the world as well.